Working minion-checkup.sls, add roster file, rm reactor.conf
Biggest hurdle, if hand-rolling roster file, was that key was in the home dir, where test state.apply was being run, so the command was finding the key, but salt didn't see it when run by salt from the state.

systemctl restart salt-minion could probably be replaced with systemctl start salt-minion but this didn't seem to make functional difference.